requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
gunicorn==21.2.0
paho-mqtt==1.6.1
python-socketio==5.11.4
python-engineio==4.10.1
